aarch64: mingw: Make long double 64 bit
On windows targets, the size of long double is 64 bits. Therefore,
unlike aarch64-**-linux-gnu, where the size of long double is 128
bits, the size of long double for aarch64-w64-mingw32 has to be 64
bits.

This commit makes changes to gcc and libgfortran so that long double is
64 bits for aarch64-w64-mingw32.

gcc/ChangeLog:

	* config/aarch64/aarch64-coff.h
	(TARGET_LONG_DOUBLE_64): Set the size of long double to 64 bits.
	(__NO_BINARY80__): Define this so that __ENABLE_BINARY80__ is
	set to 0.
	* config/aarch64/aarch64.cc
	(aarch64_scalar_mode_supported_p): Make long double 64 bits.
	(aarch64_c_mode_for_floating_type): Make long double 64 bits.

libgfortran/ChangeLog:

	* configure: Regenerate.
	* configure.ac: Add support for mingw and cygwin.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/mingw/mingw.exp: New test.
	* gcc.target/aarch64/mingw/long_double_size.c: New test.

co-authored-by: Radek Barton &lt;<A HREF="mailto:radek.barton@microsoft.com">radek.barton@microsoft.com</A>&gt;