Support experimental SARIF JSON-RPC notifications on MinGW host
[PATCH] diagnostics: add experimental SARIF JSON-RPC notifications for IDEs [PR115970]
https://gcc.gnu.org/git/?p=gcc.git;a=commitdiff;h=8a155c170b443e332b3db4922f0f0d588a17604f

introduces a regression when building the toolchain for the
aarch64-w64-mingw32 host, at least. It is most likely the same for *-*-mingw32
targets.

This patch resolves the issue when using sockets for SARIF JSON-RPC notifications
on MinGW host and adds the required socket library to the host file for *-*-mingw.

Signed-off-by: Evgeny Karpov <evgeny@kmaps.co>